Moved vulkan info class into it's own source file for easier reading
